Moved IMOD alignment methods to class file.
